Moving EditorViewState to contexts module to detangle imports
