+++ frontend/agents/AgentEditButton.js
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import { Button } from "@chakra-ui/react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPencil } from "@fortawesome/free-solid-svg-icons";
-import { EditorViewState } from "chains/ChainEditorView";
+import { EditorViewState } from "chains/editor/contexts";
 
 export const AgentEditButton = ({ agent, ...props }) => {
   const editor = React.useContext(EditorViewState);
